requirements.txt
# AI知识助手依赖包

# 深度学习框架
torch>=2.0.0
transformers>=4.36.0

# 图数据库
neo4j>=5.14.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# API框架
flask>=3.0.0
flask-cors>=4.0.0

# 配置管理
pyyaml>=6.0.1

# 日志和监控
python-json-logger>=2.0.7

# 数据验证
pydantic>=2.0.0

# 工具库
tqdm>=4.66.0
requests>=2.31.0

# 开发工具
pytest>=7.4.0
black>=23.12.0
flake8>=7.0.0

# Optional: 性能优化
# accelerate>=0.25.0
# bitsandbytes>=0.41.0  # 用于模型量化

# Optional: 向量数据库（如果需要）
# faiss-cpu>=1.7.4
# chromadb>=0.4.0

